+++ R/trajectoryConvergencePlot.R
@@ -25,7 +25,7 @@
 #' @author Miquel De \enc{Cáceres}{Caceres}, CREAF
 #' 
 #' @references
-#' Djeghri et al. (in preparation) Uncovering the relative movements of ecological trajectories.
+#' Djeghri et al. (under review) Uncovering the relative movements of ecological trajectories.
 #' 
 #' @seealso \code{\link{trajectoryConvergence}},\code{\link{trajectoryRMA}}, \code{\link{cycleShiftArrows}}
 #'