+++ R/trajectoryCyclical.R
@@ -95,7 +95,7 @@
 #' @author Miquel De \enc{Cáceres}{Caceres}, CREAF
 #' 
 #' @references
-#' Djeghri et al. (under review) Going round in cycles, but going somewhere: Ecological Trajectory Analysis as a tool to decipher seasonality and other cyclical dynamics.
+#' Djeghri et al. (2026) Going round in cycles, but going somewhere: Ecological Trajectory Analysis as a tool to decipher seasonality and other cyclical dynamics. Ecological Monographs (accepted).
 #' 
 #' @seealso \code{\link{trajectoryCyclicalPlots}}, \code{\link{cycleShiftArrows}}, \code{\link{trajectoryMetrics}}, \code{\link{trajectoryComparison}}, \code{\link{trajectoryConvergencePlot}}
 #'