+++ R/trajectoryRMAPlot.R
@@ -36,7 +36,7 @@
 #' @author Miquel De \enc{Cáceres}{Caceres}, CREAF
 #' 
 #' @references
-#' Djeghri et al. (in preparation) Uncovering the relative movements of ecological trajectories.
+#' Djeghri et al. (under review) Uncovering the relative movements of ecological trajectories.
 #'
 #' @seealso \code{\link{trajectoryRMA}}, \code{\link{trajectoryConvergencePlot}} 
 #'