+++ R/transformTrajectories.R
@@ -17,19 +17,21 @@
 #' \itemize{
 #' \item{Function \code{smoothTrajectories} performs multivariate smoothing on trajectory data using a Gaussian kernel.}
 #' \item{Function \code{centerTrajectories} shifts all trajectories to the center of the multivariate space and returns a modified distance matrix.}
+#' \item{Function \code{averageTrajectories} returns an "average" trajectory where the position of each observation is the average of the position of the corresponding observations of the input trajectories. }
 #' \item{Function \code{interpolateTrajectories} relocates trajectory ecological states to those corresponding to input times, via interpolation.}
 #' }
 #'  
 #' 
 #' @encoding UTF-8
 #' @name transformTrajectories
-#' @aliases centerTrajectories smoothTrajectories interpolateTrajectories
+#' @aliases centerTrajectories smoothTrajectories interpolateTrajectories averageTrajectories
 #' 
 #' @param x An object of class \code{\link{trajectories}}.
 #' 
 #' @details 
 #' Details of calculations are given in De \enc{Cáceres}{Caceres} et al (2019). 
-#' Function \code{centerTrajectories} performs centering of trajectories using matrix algebra as explained in Anderson (2017).
+#' Functions \code{centerTrajectories} and \code{averageTrajectories} perform centering/averaging of trajectories using matrix algebra as explained in Anderson (2017).
+#' Function \code{averageTrajectories} requires synchronous input trajectories.
 #'
 #' @return 
 #' A modified object of class \code{\link{trajectories}}, where distance matrix has been transformed. When calling \code{interpolateTrajectories}, also the
@@ -117,7 +119,7 @@ smoothTrajectories<-function(x, survey_times = NULL, kernel_scale = 1, fixed_end
 }
 
 #' @rdname transformTrajectories
-#' @param exclude An integer vector indicating sites that are excluded from trajectory centroid computation. Note: for objects of class \code{\link{cycles}}, \code{external} are excluded by default.
+#' @param exclude An integer vector indicating observations that are excluded from trajectory centroid computation. Note: for objects of class \code{\link{cycles}}, \code{external} are excluded by default.
 #' @export
 centerTrajectories<-function(x, exclude = integer(0)) {
   if(!inherits(x, "trajectories")) stop("'x' should be of class `trajectories`")
@@ -143,7 +145,7 @@ centerTrajectories<-function(x, exclude = integer(0)) {
     if(min(exclude)<1) stop("`exclude` contains values outside range")
     s_non_excluded <- sites[-exclude] 
     for(s in unique(sites)) {
-      if(sum(s_non_excluded==s)==0) stop("`exclude` cannot include all sites of a trajectory")
+      if(sum(s_non_excluded==s)==0) stop("`exclude` cannot include all observations of a trajectory")
     }
   }
   
@@ -161,14 +163,14 @@ centerTrajectories<-function(x, exclude = integer(0)) {
   H <- M%*%MASS::ginv(t(M)%*%M)%*%t(M)
   if(length(exclude)>0) {
     non_exclude <- (1:length(sites))[-exclude]
-    #Copy projection values from non-excluded site of the trajectory that the external site belongs to
+    #Copy projection values from non-excluded observations of the trajectory that the external observation belongs to
     for(i in 1:length(exclude)) {
       s <- sites[exclude[i]]
       copy_from <- which((sites==s)&((1:n)%in%non_exclude))[1]
       H[,exclude[i]] <- H[,copy_from] # Copies values for centroid removal
     }
   }
-  #Residual G matrix (when there are no excluded sites, the H matrix is symmetrical)
+  #Residual G matrix (when there are no excluded observations, the H matrix is symmetrical)
   R <- (I-t(H))%*%G%*%(I-H)
   #Backtransform to distances
   dcent<-matrix(0,n,n)
@@ -183,6 +185,88 @@ centerTrajectories<-function(x, exclude = integer(0)) {
   }
   # Replace distance matrix by dcent
   x$d <- as.dist(dcent)
+  return(x)
+}
+
+#' @rdname transformTrajectories
+#' @param group Character vector of the sites (trajectories) to be averaged. If \code{NULL} all trajectories contribute to the average  
+#' @param keep_members Boolean flag to keep the group member trajectories in the result
+#' @param output_name A string with the name for the average trajectory
+#' @export
+averageTrajectories<-function(x, group = NULL, keep_members = FALSE, output_name = "average") {
+  if(!inherits(x, "trajectories")) stop("'x' should be of class `trajectories`")
+  if(!is.synchronous(x)) stop("Trajectories need to be synchronous for trajectory averaging.")
+  
+  d <- x$d
+  surveys <- x$metadata$surveys
+  if(inherits(x, "fd.trajectories")) {
+    sites <- x$metadata$fdT
+  } else if(inherits(x, "cycles")) {
+    sites <- x$metadata$cycles
+  } else if(inherits(x, "sections")) {
+    sites <- x$metadata$sections
+  } else {
+    sites <- x$metadata$sites
+  }
+  # Check group definition
+  if(is.null(group)) {
+    group <- unique(sites)
+  } else {
+    group <- unique(as.character(group))
+  }
+  if(length(group)==0) stop("Group should be either NULL or with length > 1")
+  if(sum(group %in% sites) < length(group)) stop("Some group members were not found in `sites`")
+  Dmat <-as.matrix(d)
+  n <- nrow(Dmat)
+  
+  #Duplicate distance matrix
+  Dmat_doubled<-matrix(0,n+n,n+n)
+  for(i in 1:n) {
+    for(j in 1:n) {
+      Dmat_doubled[i,j] = Dmat[i,j]
+      Dmat_doubled[i+n,j] = Dmat[i,j]
+      Dmat_doubled[i,j+n] = Dmat[i,j]
+      Dmat_doubled[i+n,j+n] = Dmat[i,j]
+    }
+  }
+  # Define clusters by leaving unclustered the first half, while the second being clustered by surveys
+  clusters_doubled <-c(1:n, n+surveys)
+  # Remove from the second half those sites that do not conform the group
+  sel <- c(rep(TRUE,n), sites %in% group)
+  Dmat_doubled <- Dmat_doubled[sel, sel]
+  clusters_doubled <- clusters_doubled[sel]
+  # Double-center
+  G_doubled <- .gowerCentered(Dmat_doubled)
+  # Build model matrix
+  df <- data.frame(a = factor(clusters_doubled))
+  M <- model.matrix(~a,df, contrasts = list(a = "contr.helmert"))
+  #Projection matrix H
+  H <- M%*%MASS::ginv(t(M)%*%M)%*%t(M)
+  #Projected G matrix
+  P <- t(H)%*%G_doubled%*%H
+  #Backtransform to distances
+  n_final <- sum(sel)
+  Dmat_final<-matrix(0,n_final,n_final)
+  for(i in 1:n_final) {
+    for(j in i:n_final) {
+      dsq <- (P[i,i]-2*P[i,j]+P[j,j])
+      if(dsq > 0) {
+        Dmat_final[i,j] = sqrt(dsq) #truncate negative squared distances
+        Dmat_final[j,i] = Dmat_final[i,j]
+      }
+    }
+  }
+  # Replace distance matrix and metadata
+  average_selection <- sites[sites %in% group] == group[1]
+  member_selection <- rep(TRUE, n)
+  if(!keep_members) member_selection <- !(sites %in% group)
+  is_selected <- c(member_selection, average_selection)
+  x$d <- as.dist(Dmat_final[is_selected, is_selected])
+  meta_ave <- x$metadata[sites %in% group, , drop = FALSE]
+  meta_ave <- meta_ave[average_selection, , drop = FALSE]
+  meta_ave$sites <- output_name
+  x$metadata <- rbind(x$metadata[member_selection, , drop = FALSE], meta_ave)
+  row.names(x$metadata) <- 1:nrow(x$metadata)
   return(x)
 }